Extract Jira tool-result passthrough policy from agent core
